Task #463 rename and move Dialog to the new package.
